step 1-4
