+++ heidrifx_diffusion2d.egg-info/SOURCES.txt
@@ -1,0 +1,12 @@
+LICENSE
+README.md
+pyproject.toml
+setup.py
+heidrifx_diffusion2d/__init__.py
+heidrifx_diffusion2d/diffusion2d.py
+heidrifx_diffusion2d/output.py
+heidrifx_diffusion2d.egg-info/PKG-INFO
+heidrifx_diffusion2d.egg-info/SOURCES.txt
+heidrifx_diffusion2d.egg-info/dependency_links.txt
+heidrifx_diffusion2d.egg-info/requires.txt
+heidrifx_diffusion2d.egg-info/top_level.txt
